+++ blosc/bitshuffle-generic.h
@@ -81,27 +81,27 @@ extern "C" {
 
 
 /* Private functions */
-BLOSC_NO_EXPORT int64_t
+BLOSC_NO_EXPORT int32_t
 bshuf_trans_byte_elem_remainder(const void* in, void* out, const size_t size,
                                 const size_t elem_size, const size_t start);
 
-BLOSC_NO_EXPORT int64_t
+BLOSC_NO_EXPORT int32_t
 bshuf_trans_byte_elem_scal(const void* in, void* out, const size_t size,
                            const size_t elem_size);
 
-BLOSC_NO_EXPORT int64_t
+BLOSC_NO_EXPORT int32_t
 bshuf_trans_bit_byte_remainder(const void* in, void* out, const size_t size,
                                const size_t elem_size, const size_t start_byte);
 
-BLOSC_NO_EXPORT int64_t
+BLOSC_NO_EXPORT int32_t
 bshuf_trans_elem(const void* in, void* out, const size_t lda,
                  const size_t ldb, const size_t elem_size);
 
-BLOSC_NO_EXPORT int64_t
+BLOSC_NO_EXPORT int32_t
 bshuf_trans_bitrow_eight(const void* in, void* out, const size_t size,
                          const size_t elem_size);
 
-BLOSC_NO_EXPORT int64_t
+BLOSC_NO_EXPORT int32_t
 bshuf_shuffle_bit_eightelem_scal(const void* in, void* out,
                                  const size_t size, const size_t elem_size);
 
@@ -124,9 +124,9 @@ bshuf_shuffle_bit_eightelem_scal(const void* in, void* out,
  *
  */
 
-BLOSC_NO_EXPORT int64_t
-bshuf_trans_bit_elem_scal(const void* in, void* out, const size_t size,
-                          const size_t elem_size, void* tmp_buf);
+BLOSC_NO_EXPORT int32_t
+bshuf_trans_bit_elem_scal(const void* in, void* out, const int32_t size,
+                          const int32_t elem_size, void* tmp_buf);
 
 /* Unshuffle bitshuffled data.
  *
@@ -149,9 +149,9 @@ bshuf_trans_bit_elem_scal(const void* in, void* out, const size_t size,
  *
  */
 
-BLOSC_NO_EXPORT int64_t
-bshuf_untrans_bit_elem_scal(const void* in, void* out, const size_t size,
-                            const size_t elem_size, void* tmp_buf);
+BLOSC_NO_EXPORT int32_t
+bshuf_untrans_bit_elem_scal(const void* in, void* out, const int32_t size,
+                            const int32_t elem_size, void* tmp_buf);
 
 
 #ifdef __cplusplus
